+++ client/src/Routers.js
@@ -1,0 +1,20 @@
+import React from "react";
+import { Routes, Route, BrowserRouter } from "react-router-dom";
+
+import Home from "./Home";
+import Registro from "./Registro";
+
+
+const Routers = () => {
+    return (
+        <BrowserRouter>
+            <Routes>
+                <Route path="/" element={<Home />} />
+                <Route path="Registro" element={<Registro />} />
+             
+            </Routes>
+        </BrowserRouter>
+    )
+}
+
+export default Routers;
