+++ client/src/index.js
@@ -5,6 +5,6 @@ import App from './App';
 ReactDOM.render(
 
     <
-    App / > ,
+        App />,
     document.getElementById('root')
 );
